Bacula: Add generation of shutdown.sh
The bacula services will execute the file /etc/bacula/shutdown.sh after the backup.

This file is supossed to shutdown the PC if it was started for the backup.
With the option services.bacula-fd.shutdownOnFinish you can enable this script.
This will shutdown the pc if no user is logged in.